% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mrth_kernels.R
\name{get_mrth_kernels}
\alias{get_mrth_kernels}
\title{Get random walk Metropolis-Rosenbluth-Teller-Hastings kernels for 1d target}
\usage{
get_mrth_kernels(target, Sigma_proposal)
}
\arguments{
\item{target}{function taking a vector as input and returning target log-density evaluation}

\item{Sigma_proposal}{variance of the Normal random walk proposal}
}
\value{
A list containing the keys
\code{single_kernel}, \code{coupled_kernel}.
}
\description{
This function takes a target (specified through its log-pdf)
and a variance for a Normal random walk proposal, and returns a list containing the keys
\code{single_kernel}, \code{coupled_kernel} corresponding to marginal
and coupled MH kernels.

The coupling is done by reflection-maximal coupling of the proposals,
and common uniform variable for the accept/reject step. For reflection-maximal
couplings, see \code{\link{rnorm_reflectionmax}}.

The returned kernels can then be used in the functions \code{\link{sample_meetingtime}} or
\code{\link{sample_coupled_chains}} or \code{\link{sample_unbiasedestimator}}.
}
