% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highdim_coupledkernel.R
\name{maxcoupling_xi_update}
\alias{maxcoupling_xi_update}
\title{xi_update}
\usage{
maxcoupling_xi_update(
  current_xi_1,
  eta_1,
  current_xi_2,
  eta_2,
  X,
  X_transpose,
  y,
  a0,
  b0,
  std_MH
)
}
\arguments{
\item{current_xi_1, current_xi_2}{current xi values (positive scalars)}

\item{eta_1, eta_2}{current eta values (vector length p)}

\item{X}{n by p matrix}

\item{X_transpose}{Pre-calculated transpose of X}

\item{y}{length n vector}

\item{a0}{positive scalar}

\item{b0}{positive scalar}

\item{std_MH}{standard deviation of log-normal MH proposal}
}
\value{
Coupled Metropolis-Hastings update of xi given eta
}
\description{
Coupled Metropolis-Hastings update of xi given eta
}
