% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_coupled_chains_and_fish.R
\name{sample_coupled_chains_and_fish}
\alias{sample_coupled_chains_and_fish}
\title{Sample unbiased estimator of asymptotic variance in CLT for Markov chain averages}
\usage{
sample_coupled_chains_and_fish(
  single_kernel,
  coupled_kernel,
  rinit,
  h,
  k = 0,
  m = 1,
  lag = 1,
  max_iterations = Inf,
  x_0 = NULL,
  natoms = 1
)
}
\arguments{
\item{single_kernel}{A list taking a state and returning a state, performing one step of a Markov kernel}

\item{coupled_kernel}{A list taking two states and returning two states, performing one step of a coupled Markov kernel;
it also returns a boolean "identical" indicating whether the two states are identical.}

\item{rinit}{A list representing the initial state of the chain, that can be given to 'single_kernel'}

\item{h}{test function h}

\item{k}{An integer at which to start computing the unbiased estimator}

\item{m}{A time horizon: the chains are sampled until the maximum between m and the meeting time}

\item{lag}{A time lag, equal to one by default}

\item{max_iterations}{A maximum number of iterations, at which to interrupt the while loop; Inf by default}

\item{x_0}{state fixed arbitrarily to define fishy function (y in the paper)}

\item{natoms}{number of fishy function evaluations to estimate per signed measure (R in the paper)}
}
\value{
A list with
\itemize{
\item uestimator 
\item uestimator_h2
\item indexselectedatoms
\item h_at_atoms
\item htilde_at_atoms
\item weight_at_atoms
\item atomcounter
\item meetingtime
\item elapsedtime 
\item cost 
\item cost_fishyestimation
}
}
\description{
Sample unbiased estimator of asymptotic variance
in the CLT for Markov chain averages.
}
