% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_upointwisefishy.R
\name{sample_unbiasedfishy}
\alias{sample_unbiasedfishy}
\title{Sample unbiased estimator of evaluation of fishy function}
\usage{
sample_unbiasedfishy(
  coupled_kernel,
  h,
  state_x,
  state_x_0,
  max_iterations = Inf
)
}
\arguments{
\item{coupled_kernel}{A list taking two states and returning two states, performing one step of a coupled Markov kernel;
it also returns a boolean "identical" indicating whether the two states are identical.}

\item{h}{test function with which we want to solve the Poisson equation}

\item{state_x}{state of Markov chain at position x}

\item{state_x_0}{state of Markov chain at position y}

\item{max_iterations}{A maximum number of iterations, at which to interrup the while loop; Inf by default}
}
\value{
A list with
\itemize{

\item meetingtime: the meeting time; equal to Inf if while loop was interrupted

\item estimator: value of the unbiased estimator of fishy function at x
}
}
\description{
Sample two Markov chains until they meet. One 
starts from x and the other from y. The output variable is an unbiased estimator
of a fishy function \eqn{x \mapsto \sum_{t\geq 0} \{ P^t h(x) - P^t h (y) \}}.
}
