% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_unbiasedvar.R
\name{sample_unbiasedvar}
\alias{sample_unbiasedvar}
\title{Sample unbiased estimator of asymptotic variance in CLT for Markov chain averages}
\usage{
sample_unbiasedvar(
  single_kernel,
  coupled_kernel,
  rinit,
  h,
  k = 0,
  m = 1,
  lag = 1,
  x_0 = NULL,
  natoms = 1
)
}
\arguments{
\item{single_kernel}{A list taking a state and returning a state, performing one step of a Markov kernel}

\item{coupled_kernel}{A list taking two states and returning two states, performing one step of a coupled Markov kernel;
it also returns a boolean "identical" indicating whether the two states are identical.}

\item{rinit}{A list representing the initial state of the chain, that can be given to 'single_kernel'}

\item{h}{test function h}

\item{k}{An integer at which to start computing the unbiased estimator}

\item{m}{A time horizon: the chains are sampled until the maximum between m and the meeting time}

\item{lag}{A time lag, equal to one by default}

\item{x_0}{state fixed arbitrarily to define fishy function (y in the paper)}

\item{natoms}{number of fishy function evaluations to estimate per signed measure (R in the paper)}
}
\value{
A list with
\itemize{
\item estimator: 
\item covariance_term:
\item varh:
\item pih:
\item cost_fishyestimation:
\item cost_umcmc:
\item cost:
}
}
\description{
Sample unbiased estimator of asymptotic variance
in the CLT for Markov chain averages.
}
