% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umcmchandbook-package.R
\docType{package}
\name{umcmchandbook-package}
\alias{umcmchandbook-package}
\alias{umcmchandbook}
\title{umcmchandbook}
\description{
Unbiased MCMC
}
\details{
This package contains scripts to reproduce the figures of a review paper on unbiased MCMC.
}
\author{
anonymous ...
}
\keyword{package}
