% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbiasedpoisson-package.R
\docType{package}
\name{unbiasedpoisson-package}
\alias{unbiasedpoisson-package}
\alias{unbiasedpoisson}
\title{unbiasedpoisson}
\description{
Unbiased Poisson
}
\details{
This package contains scripts to reproduce the figures of a paper.
}
\author{
anonymous ...
}
\keyword{package}
