% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highdim_kernel.R
\name{xi_update}
\alias{xi_update}
\title{xi_update}
\usage{
xi_update(xi, eta, X, X_transpose, y, a0, b0, std_MH)
}
\arguments{
\item{xi}{positive scalar}

\item{eta}{length p vector}

\item{X}{n by p matrix}

\item{X_transpose}{Pre-calculated transpose of X}

\item{y}{length n vector}

\item{a0}{positive scalar}

\item{b0}{positive scalar}

\item{std_MH}{standard deviation of log-normal MH proposal}
}
\value{
Metropolis-Hastings update of xi given eta
}
\description{
Metropolis-Hastings update of xi
}
